Move these to an actually sensible place.
