+++ src/main/java/com/github/chainmailstudios/astromine/common/conveyor/ConveyorType.java
@@ -1,4 +1,4 @@
-package com.github.chainmailstudios.astromine.common.block.conveyor.interfaces;
+package com.github.chainmailstudios.astromine.common.conveyor;
 
 import net.minecraft.util.StringIdentifiable;
